Move and rename analytics-lib
