+++ packages/client/src/modules/ledger/epics.ts
@@ -7,7 +7,6 @@ import {
   wait,
 } from "@anthem/utils";
 import { LEDGER_ERRORS } from "constants/ledger-errors";
-import Analytics from "lib/analytics-lib";
 import { EpicSignature, ReduxActionTypes } from "modules/root";
 import { i18nSelector } from "modules/settings/selectors";
 import { combineEpics } from "redux-observable";
@@ -21,6 +20,7 @@ import {
   takeUntil,
   tap,
 } from "rxjs/operators";
+import Analytics from "tools/analytics-utils";
 import {
   capitalizeString,
   getQueryParamsFromUrl,
