+++ packages/client/src/ui/NetworkSelect.tsx
@@ -2,13 +2,13 @@ import { AVAILABLE_NETWORKS, NETWORK_NAME } from "@anthem/utils";
 import { Card } from "@blueprintjs/core";
 import { NetworkLogoIcon } from "assets/images";
 import { COLORS } from "constants/colors";
-import Analytics from "lib/analytics-lib";
 import { SIGNIN_TYPE } from "modules/ledger/actions";
 import Modules, { ReduxStoreState } from "modules/root";
 import { i18nSelector } from "modules/settings/selectors";
 import React from "react";
 import { connect } from "react-redux";
 import styled, { CSSProperties } from "styled-components";
+import Analytics from "tools/analytics-utils";
 import { capitalizeString, onPath } from "tools/client-utils";
 import { composeWithProps } from "tools/context-utils";
 import { View } from "./SharedComponents";
