+++ packages/client/src/ui/pages/DashboardPage.tsx
@@ -15,13 +15,13 @@ import {
 import axios from "axios";
 import { COLORS } from "constants/colors";
 import { History } from "history";
-import Analytics from "lib/analytics-lib";
 import Modules, { ReduxStoreState } from "modules/root";
 import { i18nSelector } from "modules/settings/selectors";
 import React from "react";
 import { connect } from "react-redux";
 import { Link, RouteComponentProps, withRouter } from "react-router-dom";
 import styled from "styled-components";
+import Analytics from "tools/analytics-utils";
 import ENV from "tools/client-env";
 import {
   getChartTabsForNetwork,
